+++ src/Algebra/Graph.hs
@@ -990,6 +990,7 @@ transpose :: Graph a -> Graph a
 transpose = foldg Empty Vertex Overlay (flip Connect)
 {-# INLINE transpose #-}
 
+-- TODO: Implement via 'induceJust' to reduce code duplication.
 -- | Construct the /induced subgraph/ of a given graph by removing the
 -- vertices that do not satisfy a given predicate.
 -- Complexity: /O(s)/ time, memory and size, assuming that the predicate takes
@@ -1010,13 +1011,15 @@ induce p = foldg Empty (\x -> if p x then Vertex x else Empty) (k Overlay) (k Co
     k f x     y     = f x y
 {-# INLINE [1] induce #-}
 
--- | Construct the /induced subgraph/ of a given graph by removing the 
--- vertices that are 'Nothing'.
+-- | Construct the /induced subgraph/ of a given graph by removing the vertices
+-- that are 'Nothing'.
 -- Complexity: /O(s)/ time, memory and size.
+--
 -- @
--- induceJust ('vertex' 'Nothing')                            == 'empty'
--- induceJust (gmap Just x)                                   == x
--- induceJust ('connect' (gmap Just x) ('vertex' 'Nothing'))  == x
+-- induceJust ('vertex' 'Nothing')                               == 'empty'
+-- induceJust ('edge' ('Just' x) 'Nothing')                        == 'vertex' x
+-- induceJust . 'fmap' 'Just'                                    == 'id'
+-- induceJust . 'fmap' (\\x -> if p x then 'Just' x else 'Nothing') == 'induce' p
 -- @
 induceJust :: Graph (Maybe a) -> Graph a
 induceJust = foldg Empty (maybe Empty Vertex) (k Overlay) (k Connect)